WIP: graph, mock, store: Update subscribe() signature, remove unsubscribe()
Leo is right, unsubscribing should take place by virtue of closing the
receiving end of a subscription's stream. This will cause send(change)
to fail. We can then catch that failure and remove the subscription
from our list.

TODO: Remove subscriptions when the receiving end of a subscription's
change event stream is gone.